refactor:  move errors to cognito folder
