+++ lib/cognito/adminUpdateUser.ts
@@ -1,6 +1,6 @@
 import { CognitoIdentityServiceProvider } from 'aws-sdk';
 import CloudcarError from '../errors/index';
-import MessageError from '../message.errors';
+import MessageError from './utils/message.errors';
 
 const cognitoClient = process.env.LOCAL
   ? new CognitoIdentityServiceProvider()
