+++ tsconfig.publish.json
@@ -3,7 +3,7 @@
     "lib/**/*"
   ],
   "exclude": [
-    "test/**/*"
+    "tests/**/*"
   ],
   "compilerOptions": {
       "target": "ES2018",
